% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dataImpute}
\alias{dataImpute}
\title{Impute Data with KNN, Central Tendency or Zero}
\usage{
dataImpute(
  data.tmp,
  exceptions = NULL,
  method = "KNN",
  central.tendency = "median",
  nb.neighbors = 5,
  add.imputation.indicators = FALSE
)
}
\description{
Impute Data with KNN, Central Tendency or Zero
}
