% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterReads}
\alias{filterReads}
\title{Filter Phyloseq Object by Sums of Reads by Samples and by Taxa}
\usage{
filterReads(physeq, min.sample.sum = 100, min.taxa.sum = 2, taxa.first = TRUE)
}
\description{
Filter Phyloseq Object by Sums of Reads by Samples and by Taxa
}
