% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterSampleData}
\alias{filterSampleData}
\title{Filter Phyloseq Object's Sample Data based on a Filter Expression}
\usage{
filterSampleData(physeq, filter.criteria = NULL)
}
\description{
Filter Phyloseq Object's Sample Data based on a Filter Expression
}
