% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filterTaxTable}
\alias{filterTaxTable}
\title{Filter Phyloseq Object's Taxonomic Table based on a Filter Expression}
\usage{
filterTaxTable(physeq, filter.criteria = NULL)
}
\description{
Filter Phyloseq Object's Taxonomic Table based on a Filter Expression
}
