% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{full_alpha_diversity}
\alias{full_alpha_diversity}
\title{Get and Plot Alpha-Diversity from Phyloseq Object}
\usage{
full_alpha_diversity(
  physeq,
  taxrank = NULL,
  fraction_id_name = NULL,
  fraction_ids = NULL,
  measure = "Shannon",
  x,
  group = NULL,
  plot_type = "boxplot",
  hover.variables = hover.variables,
  color_vector = c("brown", "darkgreen", "orange", "violet"),
  stat = FALSE,
  check_depth = FALSE,
  size = 1.5,
  alpha = 1
)
}
\description{
Get and Plot Alpha-Diversity from Phyloseq Object
}
