% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{full_beta_dispersion}
\alias{full_beta_dispersion}
\title{Get and Plot Beta-Diversity from Phyloseq Object}
\usage{
full_beta_dispersion(
  physeq,
  taxrank = NULL,
  fraction_id_name = NULL,
  fraction_ids = NULL,
  method = "PCoA",
  model = NULL,
  dist = "bray",
  group = NULL,
  stat = "none",
  species = FALSE,
  axis_x = 1,
  axis_y = 2,
  ...
)
}
\description{
Get and Plot Beta-Diversity from Phyloseq Object
}
