% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{get_igseq_richness}
\alias{get_igseq_richness}
\title{Get Richness Based on Ig Score Significance}
\usage{
get_igseq_richness(
  ps_ig_score,
  proportions = FALSE,
  low_lim = -5,
  high_lim = 5
)
}
\description{
Get Richness Based on Ig Score Significance
}
