% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{get_phylo_heatmap}
\alias{get_phylo_heatmap}
\title{Prepare Data for Heatmap from Phyloseq Object}
\usage{
get_phylo_heatmap(
  physeq = physeq,
  fraction_id_name = NULL,
  fraction_ids = NULL,
  taxrank.for.heatmap = NULL,
  taxrank.for.hclust = NULL,
  transform.abundances = "identity",
  distance = "bray",
  vars.to.remove.na = c()
)
}
\description{
Prepare Data for Heatmap from Phyloseq Object
}
