% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getFilterExpression}
\alias{getFilterExpression}
\title{Generate a vector of strings, each element of which is a logical
expression corresponding to a filter passed to parameters as a list}
\usage{
getFilterExpression(data, filter.criteria)
}
\description{
Generate a vector of strings, each element of which is a logical
expression corresponding to a filter passed to parameters as a list
}
