% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{ggplot_beta_dispersion}
\alias{ggplot_beta_dispersion}
\title{Ggplot Beta-Diversity Ordination}
\usage{
ggplot_beta_dispersion(
  beta.dispersion.fit,
  hover.variables = NULL,
  scaling = 1,
  comp = c(1, 2),
  label.name = NULL,
  label.levels = NULL,
  facet.name = NULL,
  facet.levels = NULL,
  facet.row.name = NULL,
  facet.row.levels = NULL,
  facet.col.name = NULL,
  facet.col.levels = NULL,
  shape.name = NULL,
  shape.levels = NULL,
  size.name = NULL,
  animation.variable.name = NULL,
  animation.variable.levels = NULL,
  remove.na.from.plot = FALSE,
  ellipses = FALSE,
  stat.beta.dispersion = NULL,
  biplot.loadings = FALSE,
  biplot.covariates = FALSE,
  arrow.labels = FALSE,
  arrow.taxonomy.labels = NULL,
  color.arrows.by.taxa = FALSE,
  arrow.cutoff.load = 0,
  arrow.cutoff.covar = 0,
  repel = FALSE,
  max.overlaps = 10,
  marginal.plot = NULL,
  raw.loadings = TRUE
)
}
\description{
Ggplot Beta-Diversity Ordination
}
