% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IG_SCORES}
\alias{IG_SCORES}
\title{Ig-Coating Score Names}
\usage{
IG_SCORES
}
\description{
Character vector of the Ig-coating score names computable via \code{\link{compute_ig_score}}
and requestable through \code{\link{getPhyloIgSeq}}'s \code{scores} argument.
}
\examples{
IG_SCORES

}
