% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.count.like}
\alias{is.count.like}
\title{Check whether abundances look like counts}
\usage{
is.count.like(x, allow.na = TRUE, consider.small.part = TRUE)
}
\description{
Check whether abundances look like counts
}
