% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_unique_taxa_table}
\alias{make_unique_taxa_table}
\title{Make taxonomy names unique at all taxonomic rank levels}
\usage{
make_unique_taxa_table(taxa_table)
}
\description{
Make taxonomy names unique at all taxonomic rank levels
}
