% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{plot_phylo_tree}
\alias{plot_phylo_tree}
\title{Plot Phylogenetic Tree from Phyloseq Object}
\usage{
plot_phylo_tree(
  physeq,
  taxrank = NULL,
  fraction_id_name = NULL,
  fraction_ids = NULL,
  circular = FALSE,
  label = NULL,
  label.levels = NULL,
  remove.na.from.plot = FALSE,
  ...
)
}
\description{
Plot Phylogenetic Tree from Phyloseq Object
}
