% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{plotReads}
\alias{plotReads}
\title{Plot Read Counts by Samples/Taxa from Phyloseq Object}
\usage{
plotReads(physeq, min.sample.sum = NA, min.taxa.sum = NA)
}
\description{
Plot Read Counts by Samples/Taxa from Phyloseq Object
}
