% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rarefy_abundances}
\alias{rarefy_abundances}
\title{Rarefy Abundances to Same Depth by multinomial resampling}
\usage{
rarefy_abundances(
  abundance_table,
  taxa_are_rows = TRUE,
  common_count_sum = NULL,
  trim_taxa = TRUE,
  trim_samples = TRUE,
  silent_warnings = FALSE
)
}
\description{
Rarefy Abundances to Same Depth by multinomial resampling
}
