% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_otu_table.R
\name{sparse_otu_table}
\alias{sparse_otu_table}
\title{Create a sparse OTU table}
\usage{
sparse_otu_table(otu)
}
\arguments{
\item{otu}{An \code{\link[phyloseq]{otu_table}} object.}
}
\value{
A \code{\link{sparse_otu_table-class}} object.
}
\description{
Converts a \code{\link[phyloseq]{otu_table}} into a
\code{\link{sparse_otu_table-class}} backed by a \code{dgCMatrix}, reducing
memory use when most counts are zero.
}
\examples{
data(ps_16s_refinement)
ot_sparse <- sparse_otu_table(phyloseq::otu_table(ps_16s_refinement))
class(ot_sparse)
dim(ot_sparse)

}
\seealso{
\code{\link{as_sparse_phyloseq}} to convert a whole
\code{phyloseq} object at once.
}
