% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{stat_beta_dispersion}
\alias{stat_beta_dispersion}
\title{Permutation Test for Beta-Diversity Ordination}
\usage{
stat_beta_dispersion(
  beta.dispersion.fit,
  facet.name = NULL,
  facet.row.name = NULL,
  facet.col.name = NULL,
  comp = c(1, 2),
  strata.name = NULL,
  label.name
)
}
\description{
Permutation Test for Beta-Diversity Ordination
}
