% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_score.R
\name{to_wider_ig_score}
\alias{to_wider_ig_score}
\title{Convert Ig Scores to Longer Format}
\usage{
to_wider_ig_score(ig_coating_agglom, scores = NULL, shared_by = NULL)
}
\description{
Convert Ig Scores to Longer Format
}
