% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_score.R
\name{agglomPhyloIgSeq}
\alias{agglomPhyloIgSeq}
\title{Agglomerate Ig Scores by Taxrank}
\usage{
agglomPhyloIgSeq(
  phyloigseq_obj,
  abundance_fraction = NULL,
  taxrank = NULL,
  make_unique_taxonomy = TRUE,
  agglom_method = NULL,
  abundance_quantile = NULL,
  min_rel_abundance = NULL
)
}
\description{
Agglomerate Ig Scores by Taxrank
}
