% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{animate_by_variable}
\alias{animate_by_variable}
\title{Animate a Ggplot Object by a Variable}
\usage{
animate_by_variable(
  ggplot.obj,
  animation.variable.name,
  return.anim = TRUE,
  save.path = NULL
)
}
\description{
Animate a Ggplot Object by a Variable
}
