% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_score.R
\name{compute_ig_score}
\alias{compute_ig_score}
\title{Compute Other Ig Scores}
\usage{
compute_ig_score(
  method = c("palm", "kau", "prob_index", "prob_ratio", "purity_corrected_prob_index",
    "purity_corrected_prob_ratio"),
  pos,
  neg = NULL,
  pre = NULL,
  ig_freq = NULL,
  pos_purity = NULL,
  neg_impurity = NULL,
  pos_fraction = NULL,
  neg_fraction = NULL
)
}
\description{
Compute Other Ig Scores
}
