% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_z.R
\name{compute_slide_z}
\alias{compute_slide_z}
\title{Compute Slide Z from MA coordinates}
\usage{
compute_slide_z(
  ma_coords,
  was_imputed = NULL,
  window_size = 50,
  empirical_null_distribution = TRUE
)
}
\description{
Compute Slide Z from MA coordinates
}
