% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{get_alpha_diversity}
\alias{get_alpha_diversity}
\title{Get Alpha-Diversity from Phyloseq Object}
\usage{
get_alpha_diversity(
  physeq,
  from_igseq = FALSE,
  transform.abundances = "identity",
  proportions = FALSE,
  low_lim = -1.96,
  high_lim = 1.96,
  taxrank = NULL,
  fraction_id_name = NULL,
  fraction_ids = NULL,
  measure = "Shannon"
)
}
\description{
Identity (raw counts) is suitable for all metrics,
}
