% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{get_beta_diversity}
\alias{get_beta_diversity}
\title{Get Beta-Diversity from Phyloseq Object}
\usage{
get_beta_diversity(
  physeq,
  taxrank = NULL,
  fraction_id_name = NULL,
  fraction_ids = NULL,
  fit.filter.name = NULL,
  fit.filter.values = NULL,
  transform.abundances = "identity",
  dist = "bray",
  method = "PCoA",
  ndims = 10,
  model = NULL,
  confounders = c(),
  species = FALSE,
  pca = TRUE,
  perplexity = NULL,
  nb.neighbors = 15,
  min.dist = 0.1
)
}
\description{
Get Beta-Diversity from Phyloseq Object
}
