% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_z.R
\name{get_ellipse_data}
\alias{get_ellipse_data}
\title{Get Confidence Ellipse Coordinates and Taxa Confidence Levels}
\usage{
get_ellipse_data(
  sorted_sample_df,
  imputed_taxa = NULL,
  empirical_null_distribution = TRUE,
  confidence_levels
)
}
\description{
Get Confidence Ellipse Coordinates and Taxa Confidence Levels
}
