% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_plot.R
\name{get_ma_coordinates}
\alias{get_ma_coordinates}
\title{Get MA Coordinates}
\usage{
get_ma_coordinates(
  sorted_sample_df,
  positive_fraction_name,
  first_negative_fraction_name,
  second_negative_fraction_name = NULL
)
}
\description{
Get MA Coordinates
}
