% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma_plot.R
\name{get_ma_plot_data}
\alias{get_ma_plot_data}
\title{Get Data for Ig-Seq MA plot}
\usage{
get_ma_plot_data(
  sorted_sample_df,
  positive_fraction_name,
  first_negative_fraction_name,
  second_negative_fraction_name = NULL,
  zero_treatments = c("keep_zeros")
)
}
\description{
Get Data for Ig-Seq MA plot
}
