% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_z.R
\name{get_slide_z}
\alias{get_slide_z}
\title{Get Slide Z score from Ig Fractions Data}
\usage{
get_slide_z(
  sorted_sample_df,
  positive_fraction_name = "pos",
  first_negative_fraction_name = "neg1",
  second_negative_fraction_name = "neg2",
  window_size = 50,
  empirical_null_distribution = TRUE,
  confidence_levels = NULL,
  imputed_taxa = NULL
)
}
\description{
Get Slide Z score from Ig Fractions Data
}
