% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igseq_data.R
\name{group_sorted_samples}
\alias{group_sorted_samples}
\title{Group Ig Fractions by Sample}
\usage{
group_sorted_samples(
  physeq,
  taxrank = NULL,
  sample_id_name,
  sample_ids = NULL,
  fraction_id_name,
  fraction_ids = NULL,
  rarefy_by_sample = TRUE,
  fractions_to_rarefy = NULL,
  transform_by_sample = "identity"
)
}
\description{
Group Ig Fractions by Sample
}
