% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_with_central_tendency}
\alias{impute_with_central_tendency}
\title{Impute with central tendency}
\usage{
impute_with_central_tendency(df, central.tendency = "median")
}
\description{
Impute with central tendency
}
