% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{PhyloIgSeq-class}
\alias{PhyloIgSeq-class}
\title{PhyloIgSeq class}
\description{
An S4 class to represent the results of an Ig-Seq experiment, including Ig coating scores,
fractions, and taxonomic or sample-level metadata.
}
\section{Slots}{

\describe{
\item{\code{ig_coating}}{A data.frame containing calculated Ig scores per taxon/sample}

\item{\code{positive_fraction_name}}{Character. Name of the positive Ig-coated fraction}

\item{\code{first_negative_fraction_name}}{Character. Name of the main negative fraction (e.g., 90\%)}

\item{\code{second_negative_fraction_name}}{Character or NULL. Name of the secondary negative fraction (e.g., 10\%)}

\item{\code{presorting_fraction_name}}{Character or NULL. Name of the pre-sorting (whole community) fraction}

\item{\code{ig_freq_name}}{Character or NULL. Name of the column containing total Ig+ frequency per sample}

\item{\code{ellipse_coords}}{A data.frame or NULL. Stores coordinates for sliding Z-score ellipses}

\item{\code{sample_data}}{A data.frame or NULL. Optional metadata for each sample}

\item{\code{tax_table}}{A data.frame or NULL. Taxonomic information}

\item{\code{phyloseq_sample_ids}}{Vector or NULL. Correspondence between phyloseq sample IDs and sample IDs used in \code{ig_coating}}

\item{\code{total_reads}}{A data.frame or NULL. Total read counts per sample and fraction before rarefaction}

\item{\code{imputed_taxa}}{List or NULL. Taxa that had zeros imputed, stored per sample}
}}

