% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{PhyloIgSeq-package}
\alias{PhyloIgSeq}
\alias{PhyloIgSeq-package}
\title{PhyloIgSeq}
\description{
Package containing wrappers mainly around \pkg{phyloseq} and \pkg{vegan} packages
for downstream analysis of 16S amplicon sequencing results and allowing to further
analyze the Ig-Seq data, computing and plotting various Ig scores. The functions
from this package are used in PhyloIgSeq Shiny Web application on
\url{https://www.immulab.fr/cms/index.php/team/tools/lab-tools}.
}
\author{
\strong{Maintainer}: Vladislav Antipin \email{Vladislav.Antipin@etu.sorbonne-universite.fr}

Authors:
\itemize{
  \item Vladislav Antipin \email{Vladislav.Antipin@etu.sorbonne-universite.fr}
}

}
