% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_score.R
\name{PhyloIgSeq_to_phyloseq}
\alias{PhyloIgSeq_to_phyloseq}
\title{Convert Ig scores from PhyloIgSeq object to phyloseq}
\usage{
PhyloIgSeq_to_phyloseq(
  phyloigseq_obj,
  score_name,
  shared_by = NULL,
  imputation_method = NULL,
  central_tendency = NULL,
  nb_neighbors = 5
)
}
\description{
Convert Ig scores from PhyloIgSeq object to phyloseq
}
