% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{plot_alpha_diversity}
\alias{plot_alpha_diversity}
\title{Plot Alpha-Diversity}
\usage{
plot_alpha_diversity(
  full.sample.data,
  hover.variables = NULL,
  x = NULL,
  x.levels = NULL,
  group = NULL,
  group.levels = NULL,
  facet.mode = "wrap",
  facet = NULL,
  facet.levels = NULL,
  facet.row = NULL,
  facet.row.levels = NULL,
  facet.col = NULL,
  facet.col.levels = NULL,
  facet_labeller = "label_value",
  shape = NULL,
  shape.levels = NULL,
  size = NULL,
  point.size = 1.5,
  remove.na.from.plot = FALSE,
  plot_type = NULL,
  color_vector = c("brown", "darkgreen", "orange", "violet"),
  stat = FALSE,
  check_depth = FALSE,
  alpha = 1
)
}
\description{
Plot Alpha-Diversity
}
