% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{plot_beta_diversity}
\alias{plot_beta_diversity}
\title{Plot Beta-Diversity Ordination}
\usage{
plot_beta_diversity(
  beta.dispersion.fit,
  hover.variables = NULL,
  scaling = 1,
  comp = c(1, 2),
  label.name = NULL,
  label.levels = NULL,
  facet.mode = "wrap",
  facet = NULL,
  facet.levels = NULL,
  facet.row = NULL,
  facet.row.levels = NULL,
  facet.col = NULL,
  facet.col.levels = NULL,
  shape.name = NULL,
  shape.levels = NULL,
  size.name = NULL,
  animation.variable.name = NULL,
  animation.variable.levels = NULL,
  remove.na.from.plot = FALSE,
  ellipses = FALSE,
  fill.ellipses = FALSE,
  stat.beta.dispersion = NULL,
  biplot.loadings = FALSE,
  biplot.covariates = FALSE,
  arrow.labels = FALSE,
  arrow.taxonomy.labels = NULL,
  color.arrows.by.taxa = FALSE,
  arrow.cutoff.load = 0,
  arrow.cutoff.covar = 0,
  repel = FALSE,
  max.overlaps = 10,
  marginal.plot = NULL,
  raw.loadings = TRUE,
  point.alpha = 1,
  point.size = 3,
  projected.alpha = 0.4,
  reverse.dim1 = FALSE,
  reverse.dim2 = FALSE
)
}
\description{
Plot Beta-Diversity Ordination
}
