% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{plot_beta_diversity}
\alias{plot_beta_diversity}
\title{Plot Beta-Diversity}
\usage{
plot_beta_diversity(
  fraction_id_name = NULL,
  fraction_ids = NULL,
  scaling = 1,
  beta.dispersion.fit,
  stat.beta.dispersion,
  axis_x = 1,
  axis_y = 2,
  nf = 5,
  type = "boxplot",
  color_vector = c("cyan4", "brown", "deepskyblue", "black", "red"),
  legend_title = NULL,
  lwd = 1,
  conf = 0.9,
  cex = 2,
  font = 2,
  pch = 20,
  draw = "lines",
  ylimits = "auto",
  xlimits = "auto",
  text = FALSE,
  ncol = 1,
  species = FALSE,
  x.intersp = 1,
  y.intersp = 0.5,
  where = "topleft",
  inset = 0.2,
  pca = TRUE,
  stat.cex = 2,
  legend.cex = 2,
  widths = c(1, 1),
  heights = c(1, 1),
  margins = c(1, 1, 1, 1),
  ...
)
}
\description{
Plot Beta-Diversity
}
