% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_score.R
\name{plot_ig_score}
\alias{plot_ig_score}
\title{Plot Any Ig score}
\usage{
plot_ig_score(
  phyloigseq_obj,
  plot_type = c("boxplot", "bubbleplot")[1],
  score_name = "slide_z",
  taxrank_score = "taxon_id",
  taxrank_facet = NULL,
  group_score = "sample_id",
  group_facet = NULL,
  score_agglom_fn = c("mean", "median")[1],
  first_score_agglom_for_each = c("sample", "taxon", "both")[1],
  z_alpha2 = 1.96,
  exclude_na = TRUE,
  transpose = FALSE,
  signif_colors = (ggsci::pal_npg())(2),
  add_stats = TRUE
)
}
\description{
Plot Any Ig score
}
