% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{plot_igseq_richness}
\alias{plot_igseq_richness}
\title{Plot Richness Based on Ig Score Significance}
\usage{
plot_igseq_richness(igseq_richness_df, group, color, exclude_ns = FALSE)
}
\description{
Plot Richness Based on Ig Score Significance
}
