% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_phylo_heatmap}
\alias{plot_phylo_heatmap}
\title{Plot ComplexHeatmap}
\usage{
plot_phylo_heatmap(
  heatmap.data,
  sort_taxa_by_diff_abundance = FALSE,
  var_for_diff_abundance = NULL,
  scale_cols = FALSE,
  nb_top_taxa = 30,
  top_annotation_vars = NULL,
  bottom_annotation_var = NULL,
  split = NULL,
  color_vector = c("white", "#88CCAA", "#771122"),
  taxa_names_par = gpar(fontsize = 15, fontface = "bold.italic"),
  label_names_par = gpar(fontsize = 15, fontface = "bold")
)
}
\description{
Plot ComplexHeatmap
}
