% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_phylo_tree}
\alias{plot_phylo_tree}
\title{Plot Phylogenetic Tree from Phyloseq Object}
\usage{
plot_phylo_tree(
  physeq,
  taxrank = NULL,
  fraction_id_name = NULL,
  fraction_ids = NULL,
  layout = "rectangular",
  tip_color = NULL,
  label.tips = NULL,
  label_size = 2.5,
  ladderize = "left",
  ...
)
}
\description{
Plot Phylogenetic Tree from Phyloseq Object
}
