% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_score.R
\name{plot_slide_z}
\alias{plot_slide_z}
\title{Plot Slide Z Score}
\usage{
plot_slide_z(
  phyloigseq_obj,
  sample_ids = NULL,
  empirical_null_distribution = TRUE,
  z_alpha2 = 1.96,
  signif_colors = c((ggsci::pal_npg())(2)[2], (ggsci::pal_npg())(2)[1]),
  ellipses = TRUE
)
}
\description{
Plot Slide Z Score
}
