% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_diversity.R
\name{stat_beta_diversity}
\alias{stat_beta_diversity}
\title{Permutation Test for Beta-Diversity Ordination}
\usage{
stat_beta_diversity(
  beta.dispersion.fit,
  facet.name = NULL,
  facet.mode = "wrap",
  facet = NULL,
  facet.row = NULL,
  facet.col = NULL,
  comp = c(1, 2),
  strata.name = NULL,
  label.name,
  pairwise = FALSE
)
}
\description{
Permutation Test for Beta-Diversity Ordination
}
